requirements.txt
PyMuPDF==1.26.7 # requires Python 3.10+
numpy==2.4.0 # requires Python 3.11+ 
pandas==2.3.3 # requires Python 3.9+
reportlab==4.4.7 # requires Python 3.9+
packaging==25.0 # requires Python 3.8+
pywin32==311 
requests==2.32.5 # requires Python 3.9+
python-dotenv==1.2.1 # requires Python 3.9+
python-docx==1.2.0 # requires Python 3.9+
matplotlib==3.10.8 # requires Python 3.10+